safari-600.7-branch/LayoutTests/fast/events/popup-blocked-from-fake-user-gesture-expected.txt

Click Here
PASS win is non-null.
PASS win is undefined.
